🐛 fix: throw on 206 path with missing destination
The else branch in `mergeIntoDestination` silently wrote a head-truncated
file when 206 was received without a pre-existing destination. The branch
is dead under current `ModelManager` callers (`resumeOffset > 0` implies
`partialURL` exists, no concurrent deletion path), but the silent
corruption would become bug-prone if #275 (cross-session resume blob
persistence) introduces new code paths.

Replace with `throw URLError(.badServerResponse)` + Logger to surface the
precondition violation explicitly. `.badServerResponse` over
`.cannotCreateFile` avoids the misleading "Cannot create file" UI copy
for what is really a protocol-state mismatch.

Add `defer { try? removeItem(at: tempURL) }` at function entry to keep
tempURL cleanup throw-safe across the new throw site and the streaming
throw sites added in the next commit (per critic Axis 4 — bundling the
deferred cleanup as a safety prerequisite for the 206-path widening).

Extract `DownloadDelegate` and `DownloadResult` to a sibling file because
the doc-comment expansion above pushes `ModelDownloader.swift` over the
400-line `file_length` cap. Mechanical move only — visibility relaxed
from `private` to internal-default for sibling-file access; no behavior
change.

Relates to #70.
